Added a git filter to convert spaces to tabs.
The .gitconfig file contains a filter named "spacestotabs" which, for
the leading spaces of every line, converts every group of 4 spaces to a
tab character. smudge applies on checkout and clean applies on checkin.

There are also settings to ensure LF is the line terminator and
whitespace issues are displayed in 'git diff'.

The .gitconfig should be included in the local git config with the
command 'git config --local include.path ../.gitconfig'.